+++ jslib/humantime/test/test.js
@@ -478,5 +478,25 @@ define(function (require) {
       console.log(a);
     });
 
+    it('`date` = 2013-03-01, `time` = 18:00:00', function () {
+      var eft = {
+          begin_at: {
+            date_word: "", 
+            date: "2013-03-01",
+            time_word: "",
+            time: "18:00:00",
+            timezone: "-05:00 EST",
+            id: 0,
+            type: "EFTime"
+          },
+          id: 0,
+          origin: "2013-03-01 13:00",
+          outputformat: 0,
+          type: "CrossTime"
+        };
+      var now = new Date();
+      var a = printEFTime(eft);
+      console.log(a);
+    });
   });
 });